+++ CHANGELOG.md
@@ -5,6 +5,9 @@ New features:
 * Context pre-initialization with TruffleRuby `--native`, which significantly
   improves startup time and loads the `did_you_mean` gem ahead of time.
 
+* The `REMOVABLE`, `MODIFIABLE` and `INSERTABLE` Truffle interop key info flags
+  have been implemented.
+
 Performance:
 
 * `Dir.glob` is much faster and more memory efficient in cases that can reduce
@@ -15,6 +18,10 @@ Performance:
 
 * `Array#dup` and `Array#shift` have been made constant-time operations by
   sharing the array storage and keeping a starting index.
+
+Bug fixes:
+
+* `Truffle::Interop.key_info` works with non-string-like names.
 
 Internal changes: